requirements.txt
openai>=1.0.0
openai-agents[voice]>=0.1.0
numpy>=1.24.0
sounddevice>=0.4.6
websocket-client>=1.7.0 